Finance_Tracker/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
fpdf2==2.7.6
openpyxl==3.1.2
python-dotenv==1.0.0